requirements.txt
python-twitter-v2==0.9.2
google-generativeai==0.3.2
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
pycoingecko==3.1.0
tweepy==4.14.0